Nuked Unification base; preparing to get rid of Unification goal
